muimaster.library: snprintf() de-linting

git-svn-id: https://svn.aros.org/svn/aros/trunk@37267 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/muimaster/classes/application.c
@@ -354,7 +354,7 @@ static IPTR Application__OM_NEW(struct IClass *cl, Object *obj, struct opSet *ms
 
         for (i = 1; i < 1000; i++)
         {
-            snprintf(portname, 255, "%s.%d", data->app_Base, i);
+            snprintf(portname, 255, "%s.%d", data->app_Base, (int)i);
             if (!find_application_by_base(cl, obj, portname))
                 break;
         }
