+++ workbench/libs/muimaster/classes/pendisplay.c
@@ -44,7 +44,7 @@ IPTR Pendisplay__OM_NEW(struct IClass *cl, Object *obj, struct opSet *msg)
     
     data = INST_DATA(cl, obj);
     /* Format identifies need to be long enforced because of RawDoFmt() limits */
-    snprintf(data->penspec.ps_buf, sizeof(data->penspec.ps_buf), "%lc%d", (LONG)PST_MUI, (LONG)MPEN_TEXT);
+    snprintf(data->penspec.ps_buf, sizeof(data->penspec.ps_buf), "%lc%d", (int)PST_MUI, (int)MPEN_TEXT);
     data->pen = -1;
     
     /* parse initial taglist */
@@ -62,7 +62,7 @@ IPTR Pendisplay__OM_NEW(struct IClass *cl, Object *obj, struct opSet *msg)
 	    	{
 		    struct MUI_RGBcolor *rgb = (struct MUI_RGBcolor *)tag->ti_Data;
 	    	    snprintf(data->penspec.ps_buf, sizeof(data->penspec.ps_buf),
-			     "%lc%08x,%08x,%08x", (LONG)PST_RGB, rgb->red, rgb->green, rgb->blue);
+			     "%lc%08x,%08x,%08x", (int)PST_RGB, (unsigned int)rgb->red, (unsigned int)rgb->green, (unsigned int)rgb->blue);
 		}
 		break;
 		
@@ -124,7 +124,7 @@ IPTR Pendisplay__OM_SET(struct IClass *cl, Object *obj, struct opSet *msg)
 		    struct MUI_RGBcolor *rgb = (struct MUI_RGBcolor *)tag->ti_Data;
 
 	    	    snprintf(data->penspec.ps_buf, sizeof(data->penspec.ps_buf),
-			     "%lc%08x,%08x,%08x", (LONG)PST_RGB, rgb->red, rgb->green, rgb->blue);
+			     "%lc%08x,%08x,%08x", (int)PST_RGB, (unsigned int)rgb->red, (unsigned int)rgb->green, (unsigned int)rgb->blue);
 		}
 		newcol = TRUE;
 		break;
@@ -296,7 +296,7 @@ IPTR Pendisplay__MUIM_Pendisplay_SetColormap(struct IClass *cl, Object *obj, str
 {
     struct MUI_PenSpec penspec;
     
-    snprintf(penspec.ps_buf, sizeof(penspec.ps_buf), "%lc%d", (LONG)PST_CMAP, msg->colormap);    
+    snprintf(penspec.ps_buf, sizeof(penspec.ps_buf), "%lc%d", (int)PST_CMAP, (int)msg->colormap);    
     set(obj, MUIA_Pendisplay_Spec, (IPTR)&penspec);
    
     return 0;
@@ -306,7 +306,7 @@ IPTR Pendisplay__MUIM_Pendisplay_SetRGB(struct IClass *cl, Object *obj, struct M
 {
     struct MUI_PenSpec penspec;
 
-    snprintf(penspec.ps_buf, sizeof(penspec.ps_buf), "%lc%08x,%08x,%08x", (LONG)PST_RGB, msg->r, msg->g, msg->b);
+    snprintf(penspec.ps_buf, sizeof(penspec.ps_buf), "%lc%08x,%08x,%08x", (int)PST_RGB, (unsigned int)msg->r, (unsigned int)msg->g, (unsigned int)msg->b);
     set(obj, MUIA_Pendisplay_Spec, (IPTR)&penspec);
    
     return 0;
@@ -316,7 +316,7 @@ IPTR Pendisplay__MUIM_Pendisplay_SetMUIPen(struct IClass *cl, Object *obj, struc
 {
     struct MUI_PenSpec penspec;
     
-    snprintf(penspec.ps_buf, sizeof(penspec.ps_buf), "%lc%d", (LONG)PST_MUI, msg->muipen);
+    snprintf(penspec.ps_buf, sizeof(penspec.ps_buf), "%lc%d", (int)PST_MUI, (int)msg->muipen);
     set(obj, MUIA_Pendisplay_Spec, (IPTR)&penspec);
    
     return 0;
